fix: updating with minor fixes and support for custom types

+++ icp/tests/python/event_plane/utils.py
@@ -18,16 +18,14 @@ import asyncio
 import logging
 import subprocess
 import time
-import uuid
 from contextlib import asynccontextmanager
 
 import pytest_asyncio
 
-from triton_distributed.icp.nats_event_plane import (
+from triton_distributed.icp import (
     DEFAULT_EVENTS_HOST,
     DEFAULT_EVENTS_PORT,
     NatsEventPlane,
-    compose_nats_url,
 )
 
 logger = logging.getLogger(__name__)
@@ -43,6 +41,7 @@ def is_port_in_use(port: int) -> bool:
 @pytest_asyncio.fixture(loop_scope="session")
 async def nats_server():
     """Fixture to start and stop a NATS server."""
+    process = None
     try:
         # Raise more intuitive error to developer if port is already in-use.
         if is_port_in_use(DEFAULT_EVENTS_PORT):
@@ -60,6 +59,7 @@ async def nats_server():
                 "-addr",
                 DEFAULT_EVENTS_HOST,
             ],
+            stdin=subprocess.DEVNULL,
             stdout=subprocess.PIPE,
             stderr=subprocess.PIPE,
         )
@@ -69,32 +69,32 @@ async def nats_server():
         logger.info("NATS server started")
         yield process
     finally:
-        logger.debug("Closing NATS server")
-
-        process.terminate()
-        # communicate() ensures we consume all stdout/stderr so they can close
-        out, err = process.communicate()
-
-        # If you want to log them:
-        logger.debug("NATS server stdout: %s", out.decode())
-        logger.debug("NATS server stderr: %s", err.decode())
-
-        if process.stdout:
-            process.stdout.close()
-        if process.stderr:
-            process.stderr.close()
-
         # Stop the NATS server
-        process.wait()
+        if process:
+            logger.debug("Closing NATS server")
+
+            process.terminate()
+            # communicate() ensures we consume all stdout/stderr so they can close
+            out, err = process.communicate()
+
+            # If you want to log them:
+            logger.debug("NATS server stdout: %s", out.decode())
+            logger.debug("NATS server stderr: %s", err.decode())
+
+            if process.stdout:
+                process.stdout.close()
+            if process.stderr:
+                process.stderr.close()
+
+            # Stop the NATS server
+            process.wait()
 
 
 @asynccontextmanager
 async def event_plane_context():
     # with nats_server_context() as server:
     print(f"Print loop plane context: {id(asyncio.get_running_loop())}")
-    server_url = compose_nats_url()
-    component_id = uuid.uuid4()
-    plane = NatsEventPlane(server_url, component_id)
+    plane = NatsEventPlane()
     await plane.connect()
     yield plane
     await plane.disconnect()
@@ -103,9 +103,7 @@ async def event_plane_context():
 @pytest_asyncio.fixture(loop_scope="function")
 async def event_plane():
     print(f"Print loop plane: {id(asyncio.get_running_loop())}")
-    server_url = compose_nats_url()
-    component_id = uuid.uuid4()
-    plane = NatsEventPlane(server_url, component_id)
+    plane = NatsEventPlane()
     await plane.connect()
     yield plane
     await plane.disconnect()
